test(apple): split apple/core index.test.ts into per-module files (#1459)
apple/core/__tests__/index.test.ts had grown to 3,487 lines / 100 test()
call sites across nine unrelated production modules with zero describe()
blocks, mirroring the aggregation disease android already fixed with its
per-concern split. Cut along the existing import boundaries: apps.test.ts,
interactions.test.ts, simulator.test.ts, physical-device-screenshot.test.ts,
screenshot.test.ts, screenshot-status-bar.test.ts, devicectl.test.ts.
config.ts and runner/runner-client.ts have no dedicated file since no test's
primary subject is one of their own exports (their symbols are only
referenced as constants/mocked dependencies inside other modules' tests).

Move-only: every test body is byte-identical to the original apart from
relocated shared setup. Device fixtures and the xcrun-stub harness used by
2+ files moved to a new apple-core-stub-helpers.ts; each split file keeps
only the vi.mock() blocks its own tests actually exercise (narrowed from
the original's blanket five-module mock, verified safe since the dropped
mocks always defaulted to real-passthrough behavior anyway).

Verified 100 test() call sites total across the new files (matches
original), and vitest.config.ts's SUBPROCESS_STUB_TESTS glob updated to
the new file list. Per repo convention, these process-spawn suites are
timeout-flaky under CPU contention; a timeout on an unrelated file is not
a regression, only assertion failures are.
