+++ src/platforms/apple/core/__tests__/apps.test.ts
@@ -12,24 +12,9 @@ vi.mock('../../../../utils/retry.ts', async (importOriginal) => {
   const actual = await importOriginal<typeof import('../../../../utils/retry.ts')>();
   return { ...actual, retryWithPolicy: vi.fn(actual.retryWithPolicy) };
 });
-vi.mock('../runner/runner-client.ts', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('../runner/runner-client.ts')>();
-  return { ...actual, runAppleRunnerCommand: vi.fn(actual.runAppleRunnerCommand) };
-});
 vi.mock('../simulator.ts', async (importOriginal) => {
   const actual = await importOriginal<typeof import('../simulator.ts')>();
-  return {
-    ...actual,
-    ensureBootedSimulator: vi.fn(actual.ensureBootedSimulator),
-    openIosSimulatorApp: vi.fn(actual.openIosSimulatorApp),
-  };
-});
-vi.mock('../screenshot-status-bar.ts', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('../screenshot-status-bar.ts')>();
-  return {
-    ...actual,
-    prepareSimulatorStatusBarForScreenshot: vi.fn(actual.prepareSimulatorStatusBarForScreenshot),
-  };
+  return { ...actual, ensureBootedSimulator: vi.fn(actual.ensureBootedSimulator) };
 });
 
 const execActual = await vi.importActual<typeof import('../../../../utils/exec.ts')>(
@@ -38,13 +23,7 @@ const execActual = await vi.importActual<typeof import('../../../../utils/exec.t
 const retryActual = await vi.importActual<typeof import('../../../../utils/retry.ts')>(
   '../../../../utils/retry.ts',
 );
-const runnerActual = await vi.importActual<typeof import('../runner/runner-client.ts')>(
-  '../runner/runner-client.ts',
-);
 const simulatorActual = await vi.importActual<typeof import('../simulator.ts')>('../simulator.ts');
-const screenshotStatusBarActual = await vi.importActual<
-  typeof import('../screenshot-status-bar.ts')
->('../screenshot-status-bar.ts');
 
 import {
   closeIosApp,
@@ -61,124 +40,32 @@ import {
 } from '../apps.ts';
 import { withMockedMacOsHelper } from './macos-helper-test-utils.ts';
 import { quitMacOsApp, resolveMacOsHelperPackageRootFrom } from '../../os/macos/helper.ts';
-import {
-  captureSimulatorScreenshotWithFallback,
-  captureSimulatorScreenshotWithRetry,
-  captureScreenshotViaRunner,
-  resolveSimulatorRunnerScreenshotCandidatePaths,
-  shouldRetryIosSimulatorScreenshot,
-} from '../screenshot.ts';
-import { shouldFallbackToRunnerForIosScreenshot } from '../physical-device-screenshot.ts';
-import { ensureBootedSimulator, openIosSimulatorApp } from '../simulator.ts';
-import {
-  invalidateSimulatorStatusBarOverrideCache,
-  prepareSimulatorStatusBarForScreenshot,
-} from '../screenshot-status-bar.ts';
-import { runAppleRunnerCommand } from '../runner/runner-client.ts';
-import { iosRunnerOverrides, performGestureApple } from '../../interactions.ts';
+import { ensureBootedSimulator } from '../simulator.ts';
 import {
   IOS_DEVICE_INSTALL_TIMEOUT_MS,
-  IOS_SIMULATOR_FOCUS_TIMEOUT_MS,
   IOS_SIMULATOR_TERMINATE_TIMEOUT_MS,
 } from '../config.ts';
 import type { DeviceInfo } from '../../../../kernel/device.ts';
-import { withDiagnosticsScope } from '../../../../utils/diagnostics.ts';
 import { AppError } from '../../../../kernel/errors.ts';
 import { runCmd } from '../../../../utils/exec.ts';
 import { retryWithPolicy } from '../../../../utils/retry.ts';
-import { parseIosDeviceAppsPayload, parseIosDeviceProcessesPayload } from '../devicectl.ts';
 import { PNG } from '../../../../utils/png.ts';
-import type { GesturePlan } from '../../../../contracts/gesture-plan.ts';
-import { requireGestureSupported } from '../../../../core/capabilities.ts';
-
-const IOS_TEST_DEVICE: DeviceInfo = {
-  platform: 'apple',
-  id: 'ios-device-1',
-  name: 'iPhone Device',
-  kind: 'device',
-  booted: true,
-};
-
-const IOS_TEST_SIMULATOR: DeviceInfo = {
-  platform: 'apple',
-  id: 'sim-1',
-  name: 'iPhone 17 Pro',
-  kind: 'simulator',
-  booted: true,
-};
-
-const MACOS_TEST_DEVICE: DeviceInfo = {
-  platform: 'apple',
-  appleOs: 'macos',
-  id: 'host-macos-local',
-  name: 'Mac',
-  kind: 'device',
-  target: 'desktop',
-  booted: true,
-};
-
-const TVOS_TEST_SIMULATOR: DeviceInfo = {
-  platform: 'apple',
-  id: 'tvos-sim-1',
-  name: 'Apple TV',
-  kind: 'simulator',
-  target: 'tv',
-  booted: true,
-};
+import {
+  IOS_TEST_DEVICE,
+  IOS_TEST_SIMULATOR,
+  MACOS_TEST_DEVICE,
+  withMockedXcrun,
+} from './apple-core-stub-helpers.ts';
 
 const mockRunCmd = vi.mocked(runCmd);
 const mockRetryWithPolicy = vi.mocked(retryWithPolicy);
-const mockRunAppleRunnerCommand = vi.mocked(runAppleRunnerCommand);
 const mockEnsureBootedSimulator = vi.mocked(ensureBootedSimulator);
-const mockOpenIosSimulatorApp = vi.mocked(openIosSimulatorApp);
-
-type MockRunCmdResult = Awaited<ReturnType<typeof runCmd>>;
-type MockRunCmdResponse = MockRunCmdResult | (() => MockRunCmdResult);
-
-const OK_RESULT: MockRunCmdResult = { exitCode: 0, stdout: '', stderr: '' };
-
-function mockRunCmdResponses(responses: Record<string, MockRunCmdResponse>): void {
-  mockRunCmd.mockImplementation(async (cmd, args) => {
-    const key = formatMockRunCmdCall(cmd, args);
-    const response = responses[key];
-    if (!response) throw new Error(`Unexpected command: ${key}`);
-    return typeof response === 'function' ? response() : response;
-  });
-}
-
-function formatMockRunCmdCall(cmd: string, args: string[]): string {
-  return `${cmd} ${args.join(' ')}`;
-}
-
-function simulatorListDevicesResult(state: string): MockRunCmdResult {
-  return {
-    exitCode: 0,
-    stdout: JSON.stringify({
-      devices: {
-        'com.apple.CoreSimulator.SimRuntime.iOS-18-6': [{ udid: 'sim-1', state }],
-      },
-    }),
-    stderr: '',
-  };
-}
-
-function simulatorStateSequence(...states: string[]): () => MockRunCmdResult {
-  let index = 0;
-  return () => simulatorListDevicesResult(states[index++] ?? states.at(-1) ?? 'Booted');
-}
-const mockPrepareStatusBarForScreenshot = vi.mocked(prepareSimulatorStatusBarForScreenshot);
 
 beforeEach(() => {
   vi.resetAllMocks();
-  invalidateSimulatorStatusBarOverrideCache(IOS_TEST_SIMULATOR);
   mockRunCmd.mockImplementation(execActual.runCmd);
   mockRetryWithPolicy.mockImplementation(retryActual.retryWithPolicy);
-  mockRunAppleRunnerCommand.mockImplementation(runnerActual.runAppleRunnerCommand);
   mockEnsureBootedSimulator.mockImplementation(simulatorActual.ensureBootedSimulator);
-  mockOpenIosSimulatorApp.mockImplementation(simulatorActual.openIosSimulatorApp);
-  mockPrepareStatusBarForScreenshot.mockImplementation(
-    screenshotStatusBarActual.prepareSimulatorStatusBarForScreenshot,
-  );
 });
 
 test('resolveMacOsHelperPackageRootFrom finds helper package from source and dist-like paths', async () => {
@@ -198,270 +85,6 @@ test('resolveMacOsHelperPackageRootFrom finds helper package from source and dis
   }
 });
 
-test('iosRunnerOverrides uses synthesized iOS coordinate taps', async () => {
-  mockRunAppleRunnerCommand.mockResolvedValue({});
-
-  const { overrides } = iosRunnerOverrides(IOS_TEST_SIMULATOR, {
-    appBundleId: 'com.example.App',
-  });
-
-  await overrides.tap(100, 200);
-  await overrides.focus(110, 210);
-
-  assert.deepEqual(mockRunAppleRunnerCommand.mock.calls[0]?.[1], {
-    command: 'tap',
-    x: 100,
-    y: 200,
-    synthesized: true,
-    appBundleId: 'com.example.App',
-  });
-  assert.deepEqual(mockRunAppleRunnerCommand.mock.calls[1]?.[1], {
-    command: 'tap',
-    x: 110,
-    y: 210,
-    synthesized: true,
-    appBundleId: 'com.example.App',
-  });
-});
-
-test('iosRunnerOverrides reads and validates the fresh gesture viewport', async () => {
-  mockRunAppleRunnerCommand.mockResolvedValue({ x: 10, y: 20, x2: 300, y2: 500 });
-  const { overrides } = iosRunnerOverrides(IOS_TEST_SIMULATOR, {
-    appBundleId: 'com.example.App',
-  });
-  assert.ok(overrides.gestureViewport);
-  assert.deepEqual(await overrides.gestureViewport(), { x: 10, y: 20, width: 300, height: 500 });
-  assert.deepEqual(mockRunAppleRunnerCommand.mock.calls[0]?.[1], {
-    command: 'gestureViewport',
-    appBundleId: 'com.example.App',
-  });
-  mockRunAppleRunnerCommand.mockResolvedValue({ x: 0, y: 0, x2: 0, y2: 500 });
-  await assert.rejects(() => overrides.gestureViewport!(), { code: 'COMMAND_FAILED' });
-});
-
-for (const [name, device] of [
-  ['macOS', MACOS_TEST_DEVICE],
-  ['tvOS', TVOS_TEST_SIMULATOR],
-] as const) {
-  test(`iosRunnerOverrides keeps ${name} coordinate taps on the standard path`, async () => {
-    mockRunAppleRunnerCommand.mockResolvedValue({});
-
-    const { overrides } = iosRunnerOverrides(device, {
-      appBundleId: 'com.example.App',
-    });
-
-    await overrides.tap(100, 200);
-
-    assert.deepEqual(mockRunAppleRunnerCommand.mock.calls[0]?.[1], {
-      command: 'tap',
-      x: 100,
-      y: 200,
-      appBundleId: 'com.example.App',
-    });
-  });
-}
-
-test('performGestureApple sends exact two-pointer pan samples through gesture', async () => {
-  mockRunAppleRunnerCommand.mockResolvedValue({ transformed: true });
-  const plan = twoFingerPanPlan();
-
-  const result = await performGestureApple(
-    IOS_TEST_SIMULATOR,
-    { appBundleId: 'com.example.App' },
-    {},
-    plan,
-  );
-
-  assert.deepEqual(result, { transformed: true });
-  assert.deepEqual(mockRunAppleRunnerCommand.mock.calls[0]?.[1], {
-    command: 'gesture',
-    gesturePlan: plan,
-    appBundleId: 'com.example.App',
-  });
-});
-
-test('Apple admission and execution share the same multi-touch refusal', async () => {
-  let admissionError: AppError | undefined;
-  try {
-    requireGestureSupported(
-      {
-        intent: 'pan',
-        origin: { x: 100, y: 200 },
-        delta: { x: 80, y: -40 },
-        pointerCount: 2,
-      },
-      IOS_TEST_DEVICE,
-    );
-  } catch (error) {
-    if (error instanceof AppError) admissionError = error;
-  }
-  assert.ok(admissionError);
-
-  await assert.rejects(
-    () => performGestureApple(IOS_TEST_DEVICE, {}, {}, twoFingerPanPlan()),
-    (error: unknown) =>
-      error instanceof AppError &&
-      error.code === 'UNSUPPORTED_OPERATION' &&
-      error.message === admissionError.message &&
-      error.details?.hint === admissionError.details?.hint,
-  );
-  assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 0);
-});
-
-test('performGestureApple composes macOS one-contact plans with the drag executor', async () => {
-  mockRunAppleRunnerCommand.mockResolvedValue({ dragged: true });
-  const plan = singlePanPlan();
-
-  await performGestureApple(MACOS_TEST_DEVICE, { appBundleId: 'com.example.App' }, {}, plan);
-
-  assert.deepEqual(mockRunAppleRunnerCommand.mock.calls[0]?.[1], {
-    command: 'drag',
-    x: 100,
-    y: 200,
-    x2: 180,
-    y2: 160,
-    durationMs: 500,
-    appBundleId: 'com.example.App',
-  });
-});
-
-test('performGestureApple composes tvOS one-contact plans with remote direction', async () => {
-  mockRunAppleRunnerCommand.mockResolvedValue({ swiped: true });
-
-  await performGestureApple(TVOS_TEST_SIMULATOR, {}, {}, singlePanPlan());
-
-  assert.deepEqual(mockRunAppleRunnerCommand.mock.calls[0]?.[1], {
-    command: 'swipe',
-    direction: 'right',
-    appBundleId: undefined,
-  });
-});
-
-test('iosRunnerOverrides maps iOS scroll to a single fused scroll command', async () => {
-  // The fused scroll resolves the frame and performs the duration-aware drag in one runner
-  // lifecycle command; no separate interactionFrame request is needed.
-  mockRunAppleRunnerCommand.mockResolvedValueOnce({
-    x: 200,
-    y: 640,
-    x2: 200,
-    y2: 160,
-    referenceWidth: 400,
-    referenceHeight: 800,
-  });
-
-  const { overrides } = iosRunnerOverrides(IOS_TEST_SIMULATOR, {
-    appBundleId: 'com.example.App',
-  });
-
-  const result = await overrides.scroll('down', { durationMs: 50 });
-
-  assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 1);
-  assert.deepEqual(mockRunAppleRunnerCommand.mock.calls[0]?.[1], {
-    command: 'scroll',
-    direction: 'down',
-    durationMs: 50,
-    appBundleId: 'com.example.App',
-  });
-  assert.deepEqual(result, {
-    x1: 200,
-    y1: 640,
-    x2: 200,
-    y2: 160,
-    referenceWidth: 400,
-    referenceHeight: 800,
-    pixels: 480,
-    durationMs: 50,
-  });
-});
-
-test('iosRunnerOverrides maps iOS scroll without duration to a fused runner scroll', async () => {
-  mockRunAppleRunnerCommand.mockResolvedValueOnce({
-    x: 200,
-    y: 640,
-    x2: 200,
-    y2: 240,
-    referenceWidth: 400,
-    referenceHeight: 800,
-  });
-
-  const { overrides } = iosRunnerOverrides(IOS_TEST_SIMULATOR, {
-    appBundleId: 'com.example.App',
-  });
-
-  await overrides.scroll('down', { pixels: 400 });
-
-  assert.deepEqual(mockRunAppleRunnerCommand.mock.calls[0]?.[1], {
-    command: 'scroll',
-    direction: 'down',
-    pixels: 400,
-    appBundleId: 'com.example.App',
-  });
-});
-
-test('iosRunnerOverrides maps tvOS scroll duration to remote press hold duration', async () => {
-  mockRunAppleRunnerCommand.mockResolvedValueOnce({
-    ok: true,
-  });
-
-  const { overrides } = iosRunnerOverrides(TVOS_TEST_SIMULATOR, {
-    appBundleId: 'com.example.App',
-  });
-
-  const result = await overrides.scroll('down', { durationMs: 50 });
-
-  assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 1);
-  assert.deepEqual(mockRunAppleRunnerCommand.mock.calls[0]?.[1], {
-    command: 'remotePress',
-    remoteButton: 'down',
-    durationMs: 50,
-    appBundleId: 'com.example.App',
-  });
-  assert.deepEqual(result, { durationMs: 50 });
-});
-
-test('iosRunnerOverrides maps macOS desktop scroll to a desktop wheel command', async () => {
-  mockRunAppleRunnerCommand.mockResolvedValueOnce({
-    x: 737.5,
-    y: 476.5,
-    referenceWidth: 400,
-    referenceHeight: 800,
-  });
-
-  const { overrides } = iosRunnerOverrides(MACOS_TEST_DEVICE, {
-    appBundleId: 'com.example.App',
-  });
-
-  const result = await overrides.scroll('down', { pixels: 200, durationMs: 50 });
-
-  assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 1);
-  assert.deepEqual(mockRunAppleRunnerCommand.mock.calls[0]?.[1], {
-    command: 'desktopScroll',
-    direction: 'down',
-    pixels: 200,
-    durationMs: 50,
-    appBundleId: 'com.example.App',
-  });
-  assert.deepEqual(result, {
-    x1: 737.5,
-    y1: 476.5,
-    referenceWidth: 400,
-    referenceHeight: 800,
-    pixels: 200,
-    durationMs: 50,
-  });
-});
-
-test('iosRunnerOverrides rejects macOS desktop scroll duration above the shared cap', async () => {
-  const { overrides } = iosRunnerOverrides(MACOS_TEST_DEVICE, {
-    appBundleId: 'com.example.App',
-  });
-
-  await assert.rejects(() => overrides.scroll('down', { pixels: 200, durationMs: 10_001 }), {
-    code: 'INVALID_ARGS',
-  });
-  assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 0);
-});
-
 test('AGENT_DEVICE_MACOS_HELPER_BIN rejects relative override paths', async () => {
   const previousHelperPath = process.env.AGENT_DEVICE_MACOS_HELPER_BIN;
   process.env.AGENT_DEVICE_MACOS_HELPER_BIN = './agent-device-macos-helper';
@@ -476,68 +99,6 @@ test('AGENT_DEVICE_MACOS_HELPER_BIN rejects relative override paths', async () =
     }
   }
 });
-
-async function withMockedXcrun(
-  tempPrefix: string,
-  script: string,
-  run: (ctx: { tmpDir: string; argsLogPath: string; device: DeviceInfo }) => Promise<void>,
-): Promise<void> {
-  const tmpDir = await fs.mkdtemp(path.join(os.tmpdir(), tempPrefix));
-  const xcrunPath = path.join(tmpDir, 'xcrun');
-  const argsLogPath = path.join(tmpDir, 'args.log');
-  const scriptWithPrivacyHelp = injectDefaultPrivacyHelp(script);
-  await fs.writeFile(xcrunPath, scriptWithPrivacyHelp, 'utf8');
-  await fs.chmod(xcrunPath, 0o755);
-
-  const previousPath = process.env.PATH;
-  const previousArgsFile = process.env.AGENT_DEVICE_TEST_ARGS_FILE;
-  process.env.PATH = `${tmpDir}${path.delimiter}${previousPath ?? ''}`;
-  process.env.AGENT_DEVICE_TEST_ARGS_FILE = argsLogPath;
-
-  try {
-    await run({ tmpDir, argsLogPath, device: IOS_TEST_DEVICE });
-  } finally {
-    process.env.PATH = previousPath;
-    if (previousArgsFile === undefined) {
-      delete process.env.AGENT_DEVICE_TEST_ARGS_FILE;
-    } else {
-      process.env.AGENT_DEVICE_TEST_ARGS_FILE = previousArgsFile;
-    }
-    await fs.rm(tmpDir, { recursive: true, force: true });
-  }
-}
-
-function injectDefaultPrivacyHelp(script: string): string {
-  if (script.includes('AGENT_DEVICE_CUSTOM_PRIVACY_HELP')) return script;
-  const helpBlock = `if [ "$1" = "simctl" ] && [ "$2" = "privacy" ] && [ "$3" = "help" ]; then
-  cat <<'HELP'
-Usage: simctl privacy <device> <action> <service> [<bundle identifier>]
-
-        service
-             The service:
-                 all - Apply the action to all services.
-                 calendar - Allow access to calendar.
-                 contacts-limited - Allow access to basic contact info.
-                 contacts - Allow access to full contact details.
-                 location - Allow access to location services when app is in use.
-                 location-always - Allow access to location services at all times.
-                 photos-add - Allow adding photos to the photo library.
-                 photos - Allow full access to the photo library.
-                 media-library - Allow access to the media library.
-                 microphone - Allow access to audio input.
-                 motion - Allow access to motion and fitness data.
-                 reminders - Allow access to reminders.
-                 siri - Allow use of the app with Siri.
-                 camera - Allow access to camera.
-                 notifications - Allow access to notifications.
-HELP
-  exit 0
-fi
-`;
-  const shebang = '#!/bin/sh\n';
-  if (!script.startsWith(shebang)) return `${shebang}${helpBlock}${script}`;
-  return `${shebang}${helpBlock}${script.slice(shebang.length)}`;
-}
 
 test('openIosApp custom scheme deep links on iOS devices require app bundle context', async () => {
   const device: DeviceInfo = {
@@ -556,721 +117,6 @@ test('openIosApp custom scheme deep links on iOS devices require app bundle cont
       return true;
     },
   );
-});
-
-test('ensureBootedSimulator opens Simulator after cold boot by default', async () => {
-  mockRunCmdResponses({
-    'xcrun simctl list devices -j': simulatorStateSequence('Shutdown', 'Booted'),
-    'xcrun simctl boot sim-1': OK_RESULT,
-    'xcrun simctl bootstatus sim-1 -b': OK_RESULT,
-    'open -a Simulator': OK_RESULT,
-  });
-
-  await ensureBootedSimulator(IOS_TEST_SIMULATOR, { focusExisting: true });
-
-  assert.equal(
-    mockRunCmd.mock.calls.some(
-      ([cmd, args]) => cmd === 'open' && args.join(' ') === '-a Simulator',
-    ),
-    true,
-  );
-});
-
-test('ensureBootedSimulator runs cold boot callback only before cold boot', async () => {
-  const onColdBootStart = vi.fn();
-  mockRunCmdResponses({
-    'xcrun simctl list devices -j': simulatorStateSequence('Shutdown', 'Booted'),
-    'xcrun simctl boot sim-1': OK_RESULT,
-    'xcrun simctl bootstatus sim-1 -b': OK_RESULT,
-    'open -a Simulator': OK_RESULT,
-  });
-
-  await ensureBootedSimulator(IOS_TEST_SIMULATOR, {
-    focusExisting: true,
-    onColdBootStart,
-  });
-
-  assert.equal(onColdBootStart.mock.calls.length, 1);
-  assert.deepEqual(onColdBootStart.mock.calls[0], [IOS_TEST_SIMULATOR]);
-});
-
-test('openIosSimulatorApp opens Simulator by default', async () => {
-  mockRunCmdResponses({
-    'open -a Simulator': OK_RESULT,
-  });
-
-  await openIosSimulatorApp();
-
-  assert.deepEqual(
-    mockRunCmd.mock.calls.map(([cmd, args]) => [cmd, args.join(' ')]),
-    [['open', '-a Simulator']],
-  );
-});
-
-test('openIosSimulatorApp uses Device Hub when opted in and falls back to Simulator', async () => {
-  mockRunCmdResponses({
-    'open -a Device Hub': {
-      exitCode: 1,
-      stdout: '',
-      stderr: 'Unable to find application named Device Hub',
-    },
-    'open -a Simulator': OK_RESULT,
-  });
-
-  await openIosSimulatorApp({ deviceHub: true });
-
-  assert.deepEqual(
-    mockRunCmd.mock.calls.map(([cmd, args]) => [cmd, args.join(' ')]),
-    [
-      ['open', '-a Device Hub'],
-      ['open', '-a Simulator'],
-    ],
-  );
-});
-
-test('ensureBootedSimulator opens Simulator when already booted by default', async () => {
-  mockRunCmdResponses({
-    'xcrun simctl list devices -j': simulatorListDevicesResult('Booted'),
-    'open -a Simulator': OK_RESULT,
-  });
-
-  await ensureBootedSimulator(IOS_TEST_SIMULATOR, { focusExisting: true });
-
-  assert.deepEqual(
-    mockRunCmd.mock.calls.map(([cmd, args]) => [cmd, args.join(' ')]),
-    [
-      ['xcrun', 'simctl list devices -j'],
-      ['open', '-a Simulator'],
-    ],
-  );
-});
-
-test('ensureBootedSimulator skips cold boot callback when already booted', async () => {
-  const onColdBootStart = vi.fn();
-  mockRunCmdResponses({
-    'xcrun simctl list devices -j': simulatorListDevicesResult('Booted'),
-    'open -a Simulator': OK_RESULT,
-  });
-
-  await ensureBootedSimulator(IOS_TEST_SIMULATOR, { focusExisting: true, onColdBootStart });
-
-  assert.equal(onColdBootStart.mock.calls.length, 0);
-});
-
-test('ensureBootedSimulator opens Device Hub without activation when already booted and opted in', async () => {
-  mockRunCmdResponses({
-    'xcrun simctl list devices -j': simulatorListDevicesResult('Booted'),
-    'open -g -a Device Hub': OK_RESULT,
-  });
-
-  await ensureBootedSimulator(IOS_TEST_SIMULATOR, { deviceHub: true, focusExisting: true });
-
-  assert.equal(
-    mockRunCmd.mock.calls.some(
-      ([cmd, args]) => cmd === 'open' && args.join(' ') === '-g -a Device Hub',
-    ),
-    true,
-  );
-  assert.equal(
-    mockRunCmd.mock.calls.some(
-      ([cmd, args]) => cmd === 'open' && args.join(' ') === '-g -a Simulator',
-    ),
-    false,
-  );
-});
-
-test('ensureBootedSimulator foregrounds Device Hub after cold boot when opted in', async () => {
-  mockRunCmdResponses({
-    'xcrun simctl list devices -j': simulatorStateSequence('Shutdown', 'Booted'),
-    'xcrun simctl boot sim-1': OK_RESULT,
-    'xcrun simctl bootstatus sim-1 -b': OK_RESULT,
-    'open -a Device Hub': OK_RESULT,
-  });
-
-  await ensureBootedSimulator(IOS_TEST_SIMULATOR, { deviceHub: true, focusExisting: true });
-
-  assert.equal(
-    mockRunCmd.mock.calls.some(
-      ([cmd, args]) => cmd === 'open' && args.join(' ') === '-a Device Hub',
-    ),
-    true,
-  );
-});
-
-test('shouldFallbackToRunnerForIosScreenshot detects removed devicectl subcommand output', () => {
-  const error = new AppError('COMMAND_FAILED', 'Failed to capture iOS screenshot', {
-    stderr: "error: Unknown option '--device'",
-  });
-  assert.equal(shouldFallbackToRunnerForIosScreenshot(error), true);
-});
-
-test('shouldFallbackToRunnerForIosScreenshot ignores unrelated command failures', () => {
-  const error = new AppError('COMMAND_FAILED', 'Failed to capture iOS screenshot', {
-    stderr: 'error: device is busy connecting',
-  });
-  assert.equal(shouldFallbackToRunnerForIosScreenshot(error), false);
-});
-
-test('shouldRetryIosSimulatorScreenshot detects simulator screen-surface timeout', () => {
-  const error = new AppError('COMMAND_FAILED', 'Detected file type from extension: PNG', {
-    stderr: 'Timeout waiting for screen surfaces',
-    exitCode: 60,
-  });
-  assert.equal(shouldRetryIosSimulatorScreenshot(error), true);
-});
-
-test('shouldRetryIosSimulatorScreenshot detects timed out simctl screenshot command', () => {
-  const error = new AppError('COMMAND_FAILED', 'xcrun timed out after 20000ms', {
-    args: ['simctl', 'io', 'sim-1', 'screenshot', '/tmp/out.png'],
-    timeoutMs: 20_000,
-  });
-  assert.equal(shouldRetryIosSimulatorScreenshot(error), true);
-});
-
-test('shouldRetryIosSimulatorScreenshot ignores unrelated screenshot failures', () => {
-  const error = new AppError('COMMAND_FAILED', 'Failed to capture iOS screenshot', {
-    stderr: 'No such file or directory',
-    exitCode: 2,
-  });
-  assert.equal(shouldRetryIosSimulatorScreenshot(error), false);
-});
-
-test('captureSimulatorScreenshotWithFallback falls back to runner after retry exhaustion', async () => {
-  const tmpDir = await fs.mkdtemp(path.join(os.tmpdir(), 'agent-device-ios-runner-fallback-'));
-  let ensureBootedCalls = 0;
-  const containerPath = path.join(tmpDir, 'container');
-  const runnerImage = path.join(containerPath, 'tmp', 'fallback.png');
-  await fs.mkdir(path.dirname(runnerImage), { recursive: true });
-  await fs.writeFile(runnerImage, 'runner-image', 'utf8');
-  mockEnsureBootedSimulator.mockImplementation(async () => {
-    ensureBootedCalls += 1;
-  });
-  mockOpenIosSimulatorApp.mockResolvedValue(undefined);
-  mockPrepareStatusBarForScreenshot.mockResolvedValue(async () => {});
-  mockRetryWithPolicy.mockRejectedValue(
-    new AppError('COMMAND_FAILED', 'Detected file type from extension: PNG', {
-      stderr: 'Timeout waiting for screen surfaces',
-      exitCode: 60,
-    }),
-  );
-  mockRunAppleRunnerCommand.mockResolvedValue({ message: 'tmp/fallback.png' });
-  mockRunCmd.mockImplementation(async (_cmd, args) => {
-    if (args.includes('get_app_container')) {
-      return { exitCode: 0, stdout: `${containerPath}\n`, stderr: '' };
-    }
-    throw new Error(`Unexpected xcrun args: ${args.join(' ')}`);
-  });
-
-  try {
-    const outPath = path.join(tmpDir, 'out.png');
-    await captureSimulatorScreenshotWithFallback(IOS_TEST_SIMULATOR, outPath, {
-      appBundleId: 'com.example.app',
-      deps: {
-        ensureBooted: ensureBootedSimulator,
-        prepareStatusBarForScreenshot: prepareSimulatorStatusBarForScreenshot,
-        captureWithRetry: captureSimulatorScreenshotWithRetry,
-        normalizeDensity: async () => {},
-        captureWithRunner: captureScreenshotViaRunner,
-        shouldFallbackToRunner: shouldRetryIosSimulatorScreenshot,
-      },
-    });
-    assert.equal(ensureBootedCalls, 1);
-    assert.equal(mockRetryWithPolicy.mock.calls.length, 1);
-    assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 1);
-    assert.equal(await fs.readFile(outPath, 'utf8'), 'runner-image');
-  } finally {
-    await fs.rm(tmpDir, { recursive: true, force: true });
-  }
-});
-
-test('captureSimulatorScreenshotWithFallback falls back to runner after simctl screenshot timeout', async () => {
-  const tmpDir = await fs.mkdtemp(path.join(os.tmpdir(), 'agent-device-ios-runner-timeout-'));
-  const containerPath = path.join(tmpDir, 'container');
-  const runnerImage = path.join(containerPath, 'tmp', 'fallback-timeout.png');
-  await fs.mkdir(path.dirname(runnerImage), { recursive: true });
-  await fs.writeFile(runnerImage, 'runner-timeout', 'utf8');
-  mockEnsureBootedSimulator.mockResolvedValue(undefined);
-  mockOpenIosSimulatorApp.mockResolvedValue(undefined);
-  mockPrepareStatusBarForScreenshot.mockResolvedValue(async () => {});
-  mockRetryWithPolicy.mockRejectedValue(
-    new AppError('COMMAND_FAILED', 'xcrun timed out after 20000ms', {
-      args: ['simctl', 'io', 'sim-1', 'screenshot', '/tmp/out.png'],
-      timeoutMs: 20_000,
-    }),
-  );
-  mockRunAppleRunnerCommand.mockResolvedValue({ message: 'tmp/fallback-timeout.png' });
-  mockRunCmd.mockImplementation(async (_cmd, args) => {
-    if (args.includes('get_app_container')) {
-      return { exitCode: 0, stdout: `${containerPath}\n`, stderr: '' };
-    }
-    throw new Error(`Unexpected xcrun args: ${args.join(' ')}`);
-  });
-
-  try {
-    const outPath = path.join(tmpDir, 'out.png');
-    await captureSimulatorScreenshotWithFallback(IOS_TEST_SIMULATOR, outPath, {
-      appBundleId: 'com.example.app',
-      deps: {
-        ensureBooted: ensureBootedSimulator,
-        prepareStatusBarForScreenshot: prepareSimulatorStatusBarForScreenshot,
-        captureWithRetry: captureSimulatorScreenshotWithRetry,
-        normalizeDensity: async () => {},
-        captureWithRunner: captureScreenshotViaRunner,
-        shouldFallbackToRunner: shouldRetryIosSimulatorScreenshot,
-      },
-    });
-    assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 1);
-    assert.equal(await fs.readFile(outPath, 'utf8'), 'runner-timeout');
-  } finally {
-    await fs.rm(tmpDir, { recursive: true, force: true });
-  }
-});
-
-test('captureSimulatorScreenshotWithFallback continues when status bar preparation fails', async () => {
-  mockPrepareStatusBarForScreenshot.mockRejectedValue(
-    new AppError('COMMAND_FAILED', 'status_bar override failed'),
-  );
-  mockEnsureBootedSimulator.mockResolvedValue(undefined);
-  mockOpenIosSimulatorApp.mockResolvedValue(undefined);
-  mockRetryWithPolicy.mockResolvedValue(undefined);
-  await captureSimulatorScreenshotWithFallback(IOS_TEST_SIMULATOR, '/tmp/out.png', {
-    appBundleId: 'com.example.app',
-    normalizeStatusBar: true,
-    deps: { normalizeDensity: async () => {} },
-  });
-  assert.equal(mockPrepareStatusBarForScreenshot.mock.calls.length, 1);
-  assert.equal(mockRetryWithPolicy.mock.calls.length > 0, true);
-  assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 0);
-});
-
-test('captureSimulatorScreenshotWithFallback can skip session-backed simulator boot probe', async () => {
-  mockEnsureBootedSimulator.mockRejectedValue(new Error('should not probe boot state'));
-  mockPrepareStatusBarForScreenshot.mockResolvedValue(async () => {});
-  mockRetryWithPolicy.mockResolvedValue(undefined);
-
-  await captureSimulatorScreenshotWithFallback(IOS_TEST_SIMULATOR, '/tmp/out.png', {
-    appBundleId: 'com.example.app',
-    skipIosSimulatorBootCheck: true,
-    deps: { normalizeDensity: async () => {} },
-  });
-
-  assert.equal(mockEnsureBootedSimulator.mock.calls.length, 0);
-  assert.equal(mockRetryWithPolicy.mock.calls.length, 1);
-  assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 0);
-});
-
-test('captureSimulatorScreenshotWithFallback boots skipped-check simulator after shutdown screenshot failure', async () => {
-  const ensureBooted = vi.fn(async () => {});
-  const prepareStatusBarForScreenshot = vi.fn(async () => async () => {});
-  let captureAttempts = 0;
-  const captureWithRetry = vi.fn(async () => {
-    captureAttempts += 1;
-    if (captureAttempts === 1) {
-      throw new AppError('COMMAND_FAILED', 'simctl screenshot failed', {
-        stderr: 'Unable to boot device in current state: Shutdown',
-        args: ['simctl', 'io', 'sim-1', 'screenshot', '/tmp/out.png'],
-      });
-    }
-  });
-  const captureWithRunner = vi.fn(async () => {});
-
-  await captureSimulatorScreenshotWithFallback(IOS_TEST_SIMULATOR, '/tmp/out.png', {
-    appBundleId: 'com.example.app',
-    skipIosSimulatorBootCheck: true,
-    deps: {
-      ensureBooted,
-      prepareStatusBarForScreenshot,
-      captureWithRetry,
-      normalizeDensity: async () => {},
-      captureWithRunner,
-      shouldFallbackToRunner: shouldRetryIosSimulatorScreenshot,
-    },
-  });
-
-  assert.equal(ensureBooted.mock.calls.length, 1);
-  assert.equal(captureWithRetry.mock.calls.length, 2);
-  assert.equal(captureWithRunner.mock.calls.length, 0);
-});
-
-test('captureSimulatorScreenshotWithFallback keeps runner fallback after skipped-check boot recovery', async () => {
-  const ensureBooted = vi.fn(async () => {});
-  const prepareStatusBarForScreenshot = vi.fn(async () => async () => {});
-  let captureAttempts = 0;
-  const captureWithRetry = vi.fn(async () => {
-    captureAttempts += 1;
-    if (captureAttempts === 1) {
-      throw new AppError('COMMAND_FAILED', 'simctl screenshot failed', {
-        stderr: 'Unable to boot device in current state: Shutdown',
-        args: ['simctl', 'io', 'sim-1', 'screenshot', '/tmp/out.png'],
-      });
-    }
-    throw new AppError('COMMAND_FAILED', 'xcrun timed out after 20000ms', {
-      args: ['simctl', 'io', 'sim-1', 'screenshot', '/tmp/out.png'],
-      timeoutMs: 20_000,
-    });
-  });
-  const captureWithRunner = vi.fn(async () => {});
-
-  await captureSimulatorScreenshotWithFallback(IOS_TEST_SIMULATOR, '/tmp/out.png', {
-    appBundleId: 'com.example.app',
-    skipIosSimulatorBootCheck: true,
-    deps: {
-      ensureBooted,
-      prepareStatusBarForScreenshot,
-      captureWithRetry,
-      normalizeDensity: async () => {},
-      captureWithRunner,
-      shouldFallbackToRunner: shouldRetryIosSimulatorScreenshot,
-    },
-  });
-
-  assert.equal(ensureBooted.mock.calls.length, 1);
-  assert.equal(captureWithRetry.mock.calls.length, 2);
-  assert.equal(captureWithRunner.mock.calls.length, 1);
-});
-
-test('captureSimulatorScreenshotWithFallback ignores status bar restore failures', async () => {
-  mockPrepareStatusBarForScreenshot.mockResolvedValue(async () => {
-    throw new AppError('COMMAND_FAILED', 'status_bar clear failed');
-  });
-  mockEnsureBootedSimulator.mockResolvedValue(undefined);
-  mockOpenIosSimulatorApp.mockResolvedValue(undefined);
-  mockRetryWithPolicy.mockResolvedValue(undefined);
-  await captureSimulatorScreenshotWithFallback(IOS_TEST_SIMULATOR, '/tmp/out.png', {
-    appBundleId: 'com.example.app',
-    normalizeStatusBar: true,
-    deps: { normalizeDensity: async () => {} },
-  });
-  assert.equal(mockPrepareStatusBarForScreenshot.mock.calls.length, 1);
-  assert.equal(mockRetryWithPolicy.mock.calls.length > 0, true);
-  assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 0);
-});
-
-test('captureSimulatorScreenshotWithFallback skips status bar normalization by default', async () => {
-  mockPrepareStatusBarForScreenshot.mockResolvedValue(async () => {});
-  mockEnsureBootedSimulator.mockResolvedValue(undefined);
-  mockOpenIosSimulatorApp.mockResolvedValue(undefined);
-  mockRetryWithPolicy.mockResolvedValue(undefined);
-
-  await captureSimulatorScreenshotWithFallback(IOS_TEST_SIMULATOR, '/tmp/out.png', {
-    appBundleId: 'com.example.app',
-    deps: { normalizeDensity: async () => {} },
-  });
-
-  assert.equal(mockPrepareStatusBarForScreenshot.mock.calls.length, 0);
-  assert.equal(mockRetryWithPolicy.mock.calls.length > 0, true);
-  assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 0);
-});
-
-test('captureSimulatorScreenshotWithFallback emits fallback diagnostic before using runner', async () => {
-  const tmpDir = await fs.mkdtemp(path.join(os.tmpdir(), 'agent-device-ios-screenshot-diag-test-'));
-  const logPath = path.join(tmpDir, 'diag.ndjson');
-  try {
-    await withDiagnosticsScope(
-      {
-        debug: true,
-        logPath,
-        session: 'ios-test',
-        requestId: 'req-1',
-        command: 'screenshot',
-      },
-      async () => {
-        const containerPath = path.join(tmpDir, 'container');
-        const runnerImage = path.join(containerPath, 'tmp', 'diag-fallback.png');
-        await fs.mkdir(path.dirname(runnerImage), { recursive: true });
-        await fs.writeFile(runnerImage, 'diag-fallback', 'utf8');
-        mockEnsureBootedSimulator.mockResolvedValue(undefined);
-        mockOpenIosSimulatorApp.mockResolvedValue(undefined);
-        mockPrepareStatusBarForScreenshot.mockResolvedValue(async () => {});
-        mockRetryWithPolicy.mockRejectedValue(
-          new AppError('COMMAND_FAILED', 'xcrun timed out after 20000ms', {
-            args: ['simctl', 'io', 'sim-1', 'screenshot', '/tmp/out.png'],
-            timeoutMs: 20_000,
-          }),
-        );
-        mockRunAppleRunnerCommand.mockResolvedValue({ message: 'tmp/diag-fallback.png' });
-        mockRunCmd.mockImplementation(async (_cmd, args) => {
-          if (args.includes('get_app_container')) {
-            return { exitCode: 0, stdout: `${containerPath}\n`, stderr: '' };
-          }
-          throw new Error(`Unexpected xcrun args: ${args.join(' ')}`);
-        });
-        await captureSimulatorScreenshotWithFallback(IOS_TEST_SIMULATOR, '/tmp/out.png', {
-          appBundleId: 'com.example.app',
-          deps: {
-            ensureBooted: ensureBootedSimulator,
-            prepareStatusBarForScreenshot: prepareSimulatorStatusBarForScreenshot,
-            captureWithRetry: captureSimulatorScreenshotWithRetry,
-            normalizeDensity: async () => {},
-            captureWithRunner: captureScreenshotViaRunner,
-            shouldFallbackToRunner: shouldRetryIosSimulatorScreenshot,
-          },
-        });
-      },
-    );
-
-    const log = await waitForFileText(logPath);
-    assert.match(log, /"phase":"ios_screenshot_fallback"/);
-    assert.match(log, /"deviceId":"sim-1"/);
-    assert.match(log, /"errorCode":"COMMAND_FAILED"/);
-    assert.match(log, /"from":"simctl_screenshot"/);
-    assert.match(log, /"to":"runner"/);
-    assert.match(log, /"commandArgs":"simctl io sim-1 screenshot \/tmp\/out\.png"/);
-  } finally {
-    await fs.rm(tmpDir, { recursive: true, force: true });
-  }
-});
-
-test('captureSimulatorScreenshotWithFallback uses simulator runner fallback by default', async () => {
-  const tmpDir = await fs.mkdtemp(path.join(os.tmpdir(), 'agent-device-ios-default-fallback-'));
-  const containerPath = path.join(tmpDir, 'container');
-  const runnerImage = path.join(containerPath, 'tmp', 'default-fallback.png');
-  await fs.mkdir(path.dirname(runnerImage), { recursive: true });
-  await fs.writeFile(runnerImage, 'default-fallback', 'utf8');
-  mockEnsureBootedSimulator.mockResolvedValue(undefined);
-  mockOpenIosSimulatorApp.mockResolvedValue(undefined);
-  mockPrepareStatusBarForScreenshot.mockResolvedValue(async () => {});
-  mockRetryWithPolicy.mockRejectedValue(
-    new AppError('COMMAND_FAILED', 'xcrun timed out after 20000ms', {
-      args: ['simctl', 'io', 'sim-1', 'screenshot', '/tmp/out.png'],
-      timeoutMs: 20_000,
-    }),
-  );
-  mockRunAppleRunnerCommand.mockResolvedValue({ message: 'tmp/default-fallback.png' });
-  mockRunCmd.mockImplementation(async (_cmd, args) => {
-    if (args.includes('get_app_container')) {
-      return { exitCode: 0, stdout: `${containerPath}\n`, stderr: '' };
-    }
-    throw new Error(`Unexpected xcrun args: ${args.join(' ')}`);
-  });
-
-  try {
-    const outPath = path.join(tmpDir, 'out.png');
-    await captureSimulatorScreenshotWithFallback(IOS_TEST_SIMULATOR, outPath, {
-      appBundleId: 'com.example.app',
-      deps: { normalizeDensity: async () => {} },
-    });
-    assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 1);
-    assert.equal(await fs.readFile(outPath, 'utf8'), 'default-fallback');
-  } finally {
-    await fs.rm(tmpDir, { recursive: true, force: true });
-  }
-});
-
-test('openIosSimulatorApp times out instead of hanging indefinitely', async () => {
-  mockRunCmd.mockImplementation(async (cmd, args, options) => {
-    assert.equal(cmd, 'open');
-    assert.deepEqual(args, ['-a', 'Simulator']);
-    assert.equal(options?.timeoutMs, IOS_SIMULATOR_FOCUS_TIMEOUT_MS);
-    throw new AppError('COMMAND_FAILED', 'open timed out after 10000ms', {
-      timeoutMs: options?.timeoutMs,
-    });
-  });
-
-  await assert.rejects(
-    () => openIosSimulatorApp(),
-    (error: unknown) => {
-      assert.equal(error instanceof AppError, true);
-      assert.equal((error as AppError).code, 'COMMAND_FAILED');
-      assert.match((error as AppError).message, /open timed out after 10000ms/);
-      return true;
-    },
-  );
-});
-
-test('prepareSimulatorStatusBarForScreenshot restores prior visible overrides', async () => {
-  await withMockedXcrun(
-    'agent-device-ios-status-bar-restore-test-',
-    `#!/bin/sh
-echo "$*" >> "$AGENT_DEVICE_TEST_ARGS_FILE"
-if [ "$1" = "simctl" ] && [ "$2" = "status_bar" ] && [ "$4" = "list" ]; then
-  cat <<'OUT'
-Current Status Bar Overrides:
-=============================
-Time: 6:07
-DataNetworkType: 0
-WiFi Mode: 2, WiFi Bars: 0
-Cell Mode: 2, Cell Bars: 0
-Operator Name: No Service
-Battery State: 1, Battery Level: 42, Not Charging: 0
-OUT
-  exit 0
-fi
-if [ "$1" = "simctl" ] && [ "$2" = "status_bar" ] && [ "$4" = "clear" ]; then
-  exit 0
-fi
-if [ "$1" = "simctl" ] && [ "$2" = "status_bar" ] && [ "$4" = "override" ]; then
-  exit 0
-fi
-echo "unexpected xcrun args: $*" >&2
-exit 1
-`,
-    async ({ argsLogPath }) => {
-      const restore = await prepareSimulatorStatusBarForScreenshot(IOS_TEST_SIMULATOR);
-      await restore();
-
-      const logLines = (await fs.readFile(argsLogPath, 'utf8')).trim().split('\n').filter(Boolean);
-      assert.deepEqual(logLines, [
-        'simctl status_bar sim-1 list',
-        'simctl status_bar sim-1 clear',
-        'simctl status_bar sim-1 override --time 9:41 --dataNetwork wifi --wifiMode active --wifiBars 3 --batteryState charged --batteryLevel 100',
-        'simctl status_bar sim-1 clear',
-        'simctl status_bar sim-1 override --dataNetwork hide --wifiMode failed --wifiBars 0 --cellularMode failed --cellularBars 0 --operatorName No Service',
-      ]);
-    },
-  );
-});
-
-test('prepareSimulatorStatusBarForScreenshot skips known redundant status bar commands', async () => {
-  await withMockedXcrun(
-    'agent-device-ios-status-bar-no-overrides-test-',
-    `#!/bin/sh
-echo "$*" >> "$AGENT_DEVICE_TEST_ARGS_FILE"
-if [ "$1" = "simctl" ] && [ "$2" = "status_bar" ] && [ "$4" = "list" ]; then
-  cat <<'OUT'
-Current Status Bar Overrides:
-=============================
-OUT
-  exit 0
-fi
-if [ "$1" = "simctl" ] && [ "$2" = "status_bar" ] && [ "$4" = "clear" ]; then
-  exit 0
-fi
-if [ "$1" = "simctl" ] && [ "$2" = "status_bar" ] && [ "$4" = "override" ]; then
-  exit 0
-fi
-echo "unexpected xcrun args: $*" >&2
-exit 1
-`,
-    async ({ argsLogPath }) => {
-      const restoreFirst = await prepareSimulatorStatusBarForScreenshot(IOS_TEST_SIMULATOR);
-      await restoreFirst();
-      const restoreSecond = await prepareSimulatorStatusBarForScreenshot(IOS_TEST_SIMULATOR);
-      await restoreSecond();
-
-      const logLines = (await fs.readFile(argsLogPath, 'utf8')).trim().split('\n').filter(Boolean);
-      assert.deepEqual(logLines, [
-        'simctl status_bar sim-1 list',
-        'simctl status_bar sim-1 override --time 9:41 --dataNetwork wifi --wifiMode active --wifiBars 3 --batteryState charged --batteryLevel 100',
-        'simctl status_bar sim-1 clear',
-        'simctl status_bar sim-1 override --time 9:41 --dataNetwork wifi --wifiMode active --wifiBars 3 --batteryState charged --batteryLevel 100',
-        'simctl status_bar sim-1 clear',
-      ]);
-    },
-  );
-});
-
-test('prepareSimulatorStatusBarForScreenshot still normalizes when snapshotting current overrides fails', async () => {
-  await withMockedXcrun(
-    'agent-device-ios-status-bar-snapshot-failure-test-',
-    `#!/bin/sh
-echo "$*" >> "$AGENT_DEVICE_TEST_ARGS_FILE"
-if [ "$1" = "simctl" ] && [ "$2" = "status_bar" ] && [ "$4" = "list" ]; then
-  echo "list failed" >&2
-  exit 1
-fi
-if [ "$1" = "simctl" ] && [ "$2" = "status_bar" ] && [ "$4" = "clear" ]; then
-  exit 0
-fi
-if [ "$1" = "simctl" ] && [ "$2" = "status_bar" ] && [ "$4" = "override" ]; then
-  exit 0
-fi
-echo "unexpected xcrun args: $*" >&2
-exit 1
-`,
-    async ({ argsLogPath }) => {
-      const restore = await prepareSimulatorStatusBarForScreenshot(IOS_TEST_SIMULATOR);
-      await restore();
-
-      const logLines = (await fs.readFile(argsLogPath, 'utf8')).trim().split('\n').filter(Boolean);
-      assert.deepEqual(logLines, [
-        'simctl status_bar sim-1 list',
-        'simctl status_bar sim-1 clear',
-        'simctl status_bar sim-1 override --time 9:41 --dataNetwork wifi --wifiMode active --wifiBars 3 --batteryState charged --batteryLevel 100',
-        'simctl status_bar sim-1 clear',
-      ]);
-    },
-  );
-});
-
-async function waitForFileText(filePath: string, attempts = 20): Promise<string> {
-  let lastError: unknown;
-  for (let attempt = 0; attempt < attempts; attempt += 1) {
-    try {
-      return await fs.readFile(filePath, 'utf8');
-    } catch (error) {
-      lastError = error;
-      await new Promise((resolve) => setTimeout(resolve, 10));
-    }
-  }
-  throw lastError;
-}
-
-test('resolveSimulatorRunnerScreenshotCandidatePaths includes tmp-based and basename fallbacks', () => {
-  const containerPath = '/tmp/container';
-  const candidates = resolveSimulatorRunnerScreenshotCandidatePaths(
-    containerPath,
-    '/var/mobile/Containers/Data/Application/abc/tmp/screenshot-1.png',
-  );
-  assert.equal(candidates.includes(path.join(containerPath, 'tmp', 'screenshot-1.png')), true);
-  assert.equal(
-    candidates.includes('/var/mobile/Containers/Data/Application/abc/tmp/screenshot-1.png'),
-    true,
-  );
-});
-
-test('resolveSimulatorRunnerScreenshotCandidatePaths handles empty runner path', () => {
-  assert.deepEqual(resolveSimulatorRunnerScreenshotCandidatePaths('/tmp/container', '   '), []);
-});
-
-test('captureScreenshotViaRunner reuses a verified simulator container path', async () => {
-  const tmpDir = await fs.mkdtemp(path.join(os.tmpdir(), 'agent-device-runner-cache-'));
-  const containerPath = path.join(tmpDir, 'container');
-  const runnerImage = path.join(containerPath, 'tmp', 'capture.png');
-  const device = { ...IOS_TEST_SIMULATOR, id: 'sim-runner-container-cache' };
-  await fs.mkdir(path.dirname(runnerImage), { recursive: true });
-  await fs.writeFile(runnerImage, 'runner-image', 'utf8');
-  mockRunAppleRunnerCommand.mockResolvedValue({ message: 'tmp/capture.png' });
-  mockRunCmd.mockImplementation(async (_cmd, args) => {
-    if (args.includes('get_app_container')) {
-      return { exitCode: 0, stdout: `${containerPath}\n`, stderr: '' };
-    }
-    throw new Error(`Unexpected xcrun args: ${args.join(' ')}`);
-  });
-
-  try {
-    const firstPath = path.join(tmpDir, 'first.png');
-    const secondPath = path.join(tmpDir, 'second.png');
-    await captureScreenshotViaRunner(device, firstPath);
-    await captureScreenshotViaRunner(device, secondPath);
-
-    assert.equal(await fs.readFile(firstPath, 'utf8'), 'runner-image');
-    assert.equal(await fs.readFile(secondPath, 'utf8'), 'runner-image');
-    assert.equal(mockRunAppleRunnerCommand.mock.calls.length, 2);
-    assert.equal(
-      mockRunCmd.mock.calls.filter(([, args]) => args.includes('get_app_container')).length,
-      1,
-    );
-  } finally {
-    await fs.rm(tmpDir, { recursive: true, force: true });
-  }
-});
-
-test('captureScreenshotViaRunner copies macOS runner screenshots from the host', async () => {
-  const tmpDir = await fs.mkdtemp(path.join(os.tmpdir(), 'agent-device-macos-screenshot-'));
-  const sourcePath = path.join(tmpDir, 'runner.png');
-  const outPath = path.join(tmpDir, 'screen.png');
-  await fs.writeFile(sourcePath, 'runner-image', 'utf8');
-  mockRunAppleRunnerCommand.mockResolvedValue({ message: sourcePath });
-  mockRunCmd.mockRejectedValue(new Error('macOS screenshot must not invoke xcrun devicectl'));
-
-  try {
-    await captureScreenshotViaRunner(MACOS_TEST_DEVICE, outPath);
-    assert.equal(await fs.readFile(outPath, 'utf8'), 'runner-image');
-    assert.equal(mockRunCmd.mock.calls.length, 0);
-  } finally {
-    await fs.rm(tmpDir, { recursive: true, force: true });
-  }
 });
 
 test('screenshotIos retries simulator capture timeouts and eventually succeeds', async () => {
@@ -2398,70 +1244,6 @@ test('pushIosNotification uses simctl push with temporary payload file', async (
   }
 });
 
-test('parseIosDeviceAppsPayload maps devicectl app entries', () => {
-  const apps = parseIosDeviceAppsPayload({
-    result: {
-      apps: [
-        {
-          bundleIdentifier: 'com.apple.Maps',
-          name: 'Maps',
-          url: 'file:///Applications/Maps.app/',
-        },
-        {
-          bundleIdentifier: 'com.example.NoName',
-        },
-      ],
-    },
-  });
-
-  assert.equal(apps.length, 2);
-  assert.deepEqual(apps[0], {
-    bundleId: 'com.apple.Maps',
-    name: 'Maps',
-    url: 'file:///Applications/Maps.app/',
-  });
-  assert.equal(apps[1]!.bundleId, 'com.example.NoName');
-  assert.equal(apps[1]!.name, 'com.example.NoName');
-  assert.equal(apps[1]!.url, undefined);
-});
-
-test('parseIosDeviceAppsPayload ignores malformed entries', () => {
-  const apps = parseIosDeviceAppsPayload({
-    result: {
-      apps: [null, {}, { name: 'Missing bundle id' }, { bundleIdentifier: '' }],
-    },
-  });
-  assert.deepEqual(apps, []);
-});
-
-test('parseIosDeviceProcessesPayload maps running process entries', () => {
-  const processes = parseIosDeviceProcessesPayload({
-    result: {
-      runningProcesses: [
-        {
-          executable: 'file:///private/var/containers/Bundle/Application/ABC123/Demo.app/Demo',
-          processIdentifier: 421,
-        },
-        {
-          executable: 'file:///usr/libexec/backboardd',
-          processIdentifier: 72,
-        },
-      ],
-    },
-  });
-
-  assert.deepEqual(processes, [
-    {
-      executable: 'file:///private/var/containers/Bundle/Application/ABC123/Demo.app/Demo',
-      pid: 421,
-    },
-    {
-      executable: 'file:///usr/libexec/backboardd',
-      pid: 72,
-    },
-  ]);
-});
-
 test('resolveIosApp resolves app display name on iOS physical devices', async () => {
   const tmpDir = await fs.mkdtemp(path.join(os.tmpdir(), 'agent-device-ios-app-resolve-'));
   const xcrunPath = path.join(tmpDir, 'xcrun');
@@ -3439,49 +2221,3 @@ exit 1
     },
   );
 });
-
-function twoFingerPanPlan(): Extract<GesturePlan, { topology: 'two' }> {
-  return {
-    topology: 'two',
-    intent: 'pan',
-    durationMs: 32,
-    viewport: { x: 0, y: 0, width: 400, height: 800 },
-    pointers: [
-      {
-        pointerId: 0,
-        samples: [
-          { offsetMs: 0, point: { x: 100, y: 80 } },
-          { offsetMs: 16, point: { x: 110, y: 85 } },
-          { offsetMs: 32, point: { x: 120, y: 90 } },
-        ],
-      },
-      {
-        pointerId: 1,
-        samples: [
-          { offsetMs: 0, point: { x: 100, y: 120 } },
-          { offsetMs: 16, point: { x: 110, y: 125 } },
-          { offsetMs: 32, point: { x: 120, y: 130 } },
-        ],
-      },
-    ],
-  };
-}
-
-function singlePanPlan(): Extract<GesturePlan, { topology: 'single' }> {
-  return {
-    topology: 'single',
-    intent: 'pan',
-    executionProfile: 'timed-pan',
-    durationMs: 500,
-    viewport: { x: 0, y: 0, width: 400, height: 800 },
-    pointers: [
-      {
-        pointerId: 0,
-        samples: [
-          { offsetMs: 0, point: { x: 100, y: 200 } },
-          { offsetMs: 500, point: { x: 180, y: 160 } },
-        ],
-      },
-    ],
-  };
-}
