+++ src/platforms/apple/core/__tests__/physical-device-screenshot.test.ts
@@ -1,0 +1,18 @@
+import { test } from 'vitest';
+import assert from 'node:assert/strict';
+import { shouldFallbackToRunnerForIosScreenshot } from '../physical-device-screenshot.ts';
+import { AppError } from '../../../../kernel/errors.ts';
+
+test('shouldFallbackToRunnerForIosScreenshot detects removed devicectl subcommand output', () => {
+  const error = new AppError('COMMAND_FAILED', 'Failed to capture iOS screenshot', {
+    stderr: "error: Unknown option '--device'",
+  });
+  assert.equal(shouldFallbackToRunnerForIosScreenshot(error), true);
+});
+
+test('shouldFallbackToRunnerForIosScreenshot ignores unrelated command failures', () => {
+  const error = new AppError('COMMAND_FAILED', 'Failed to capture iOS screenshot', {
+    stderr: 'error: device is busy connecting',
+  });
+  assert.equal(shouldFallbackToRunnerForIosScreenshot(error), false);
+});
